% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmv_helper_compensation.R
\name{cmv_helper_compensation}
\alias{cmv_helper_compensation}
\title{CMV compensation helper function}
\usage{
cmv_helper_compensation(
  ff_unapplied_compensations,
  compensation_type = c("manual", "auto", "auto_singlestain"),
  expected_max_proportion_value = 5,
  original_spillover_keyword = "$SPILLOVER"
)
}
\arguments{
\item{ff_unapplied_compensations}{A flowFrame object with the unapplied compensations, generated for the CMV data from Glehr et al. \link{TO BE SUBMITTED}}

\item{compensation_type}{What compensation to use. Default is "manual". Other options are "auto" and "auto_singlestain".
"auto" calculated the spillover matrix and autofluorescence vector  based on the DuraClone
Compensation TUBES from Beckman Coulter. "auto_singlestain" calculates the spillover matrix and
autofluorescence vector  based on the single stain controls from the CMV data.
"manual" is the default option, which uses the manually curated spillover matrix and autofluorescence vector.}

\item{expected_max_proportion_value}{The expected maximum value for the spillover matrix and autofluorescence vector. Default is 1 which corresponds to 100\%. However, technically nothing speaks against spillover values above 100\%.
Within Kaluza, the spillover values can be a maximum of +/- 500\% (+/-5). Thus the default value of 5. This is a safety measure to avoid using the already calculated absolute autofluorescence values.}

\item{original_spillover_keyword}{The keyword to use for the original spillover matrix. Default is "$SPILLOVER".}
}
\description{
This function is a helper function to apply the compensation to the CMV data from
Glehr et al. \link{TO BE SUBMITTED}. It reads the spillover matrix and autofluorescence vector from
the FCS file and applies the compensation to the data. Different compensation options are stored
within the FCS files and can hereby be selected and applied easily.
}
