% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_rescale.R
\name{create_mapping}
\alias{create_mapping}
\title{Creating a mapping for named mfis and read files}
\usage{
create_mapping(
  named_mfis,
  sample_names,
  FUN.sample_names_to_mapping_df = sample_names_to_mapping_df_packageexample,
  types_single_staining = c("single_CD3", "single_staining")
)
}
\arguments{
\item{named_mfis}{A named list of MFI files, one element should look like this:

$Align_02.Lot_A.EX
FL1 CD3 FITC FL2 CD3 PE FL3 CD3 ECD FL4 CD3 PC5.5 FL5 CD3 PC7 FL6 CD3 APC FL7 CD3 AF700 FL8 CD3 AA750 FL9 CD3 PB FL10 CD3 KrO
MFI POSITIVE POPULATION                32569.91  110339.02   127206.20     532379.81   626938.75   203029.61      61649.41      48720.48   18348.22      7848.84
MFI NEGATIVE POPULATION  UNSTAINED       236.02     213.52      214.93        231.15      190.11      337.84        219.20        234.48     342.80       636.98

where the names consists of \if{html}{\out{<sample>}}.\if{html}{\out{<lot>}}.\if{html}{\out{<device>}}}
}
\value{
A tibble with the following columns:
\itemize{
\item mfi_file: The name of the MFI file
\item negative: The name of the negative control
\item sample: The name of the sample which should be rescaled with the mfi_file
\item lot: The lot of the MFI-sample
\item device: The device used for the sample
\item lot_verify: The lot of the verification sample
}
}
\description{
This is a very specific function to create a mapping between the MFI files and
the read files for our usecase.
}
