% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytobench-package.R
\docType{package}
\name{cytobench-package}
\alias{cytobench}
\alias{cytobench-package}
\title{cytobench: Bench of Tools for Cytometry Data and Relativisation}
\description{
cytobench provides a comprehensive toolkit for preprocessing, normalization, and analysis of flow cytometry data. Especially designed for performing relativisation of flow cytometry data, FlowSOM convenience functions and further convenience functions for flow cytometry data analysis.
}
\author{
\strong{Maintainer}: Gunther Glehr \email{gunthergl@gmx.net} (\href{https://orcid.org/0000-0002-1495-9162}{ORCID})

}
\keyword{internal}
