% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{example_processed}
\alias{example_processed}
\title{Load and Preprocess Example FCS File}
\usage{
example_processed(ncells = 250)
}
\arguments{
\item{ncells}{Integer.
Number of cells to use from system.file("extdata", "68983.fcs", package = "FlowSOM")}
}
\value{
A \code{flowFrame} object containing compensated and transformed cytometry data.
}
\description{
Loads an example FCS file provided by the \code{FlowSOM} package and applies
compensation and logicle transformation to prepare the data for downstream analysis.
}
\details{
This function demonstrates a minimal preprocessing pipeline using \code{flowCore}
and \code{flowWorkspace}, including compensation using the embedded spillover matrix
and a logicle transformation based on the compensated channels.
}
\examples{
ff <- example_processed()
flowCore::exprs(ff)[1:5, ]

}
\keyword{test-helper}
