% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_fcs.R
\name{export_fcs}
\alias{export_fcs}
\title{Export fcs files}
\usage{
export_fcs(
  matrix_list,
  safety_scaling = 1.2,
  safety_shift = 0,
  outdir = ".",
  use.names = FALSE,
  verbose = TRUE,
  new_colnames_to = c("description", "colnames"),
  new_colnames = 1,
  extreme_template = NULL
)
}
\arguments{
\item{matrix_list}{List of matrices, should all have the same number of columns.}

\item{safety_scaling}{An .fcs file is generated from the maximum cells (high and low). This safety scaling
is soleily for the range of the .fcs files. This does not change any cell values.}

\item{safety_shift}{CAREFUL! Introducing a value here is changing all cell values. It adds this value
to all cells. This is "necessary" as Kaluza cannot visualize negative values on a
linear scale in .fcs files.}

\item{outdir}{Where should the list of matrices be written to. Filenames come from names(matrix_list).}

\item{use.names}{Should the column names of each matrix be used when concattenating
the extreme .fcs?}

\item{verbose}{Verbose output?}

\item{new_colnames_to}{In our usecase, there is often the case where we have:

Sample1:
FL1 CD3 FITC,FL2 CD3 PE,FL3 CD3 ECD,FL4 CD3 PC5.5,FL5 CD3 PC7,FL6 CD3 APC,FL7 CD3 AF700,FL8 CD3 AA750,FL9 CD3 PB,FL10 CD3 KrO
Sample2:
FL1 CD3 FITC,FL2 empty,FL3 empty,FL4 empty,FL5 empty,FL6 empty,FL7 empty,FL8 empty,FL9 empty,FL10 empty

In this case, we want to use the column names of Sample1, but we want to retain the
column names of Sample2. The original column names of each sample are saved in either
the description or the colnames of the .fcs file. \code{Colnames} are the relevant
within Kaluza where the marker names are extracted (and gates applied).}

\item{new_colnames}{Can be single element numeric or character - referring to the
sample whose column names should be used.
Alternatively, give a named vector of strings with the same length as the matrix has columns, e.g.:

c(
"FSC-A" = NA,
"SSC-A" = NA,
"FL1-A" = "marker",
"FL4-A" = "marker",
"FL5-A" = "marker",
"FL7-A" = "marker",
"FL8-A" = "marker",
"FL9-A" = "marker",
"FL10-A" = "marker",
"FL11-A" = "marker",
"FL12-A" = "marker",
"FL13-A" = "marker",
"FL14-A" = "marker",
"FL16-A" = "marker",
"FL21-A" = "marker",
"FSC-Width" = NA,
"TIME" = NA
)

Every value which is called "marker" will be used as a marker name setting
flowCore::markernames() IN ORDER! The other values are ignored.}

\item{extreme_template}{If you have an extreme sample which you want to use as a template for all other samples,
you can provide it here. Usually, this would be the returned fcs_extreme_copy from a
previous run of this function.
Essentially, it needs at least 2 cells (high and low) as "extreme" values.
The cells are replaced by the actual cells given in the matrix_list.}
}
\value{
fcs_extreme_copy:
A copy of the extreme_template which was used to save all samples. The fcs-file
ranges for all samples come from this extreme sample.
}
\description{
Export a list of matrices, usually \code{data.table}s.
}
