% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_mfi.R
\name{extract_relevant_mfis_multistain}
\alias{extract_relevant_mfis_multistain}
\title{Extract Multi-Stain Median Fluorescence Intensity (MFI)
This function extracts the median fluorescence intensity (MFI) from multi-stained FCS files.}
\usage{
extract_relevant_mfis_multistain(
  loaded_fcs_multistain,
  transform_fun = function(x) {
     asinh(x/1000)
 },
  relevant_columns,
  seed = 42
)
}
\arguments{
\item{loaded_fcs_multistain}{A list of cytosets containing the loaded FCS files.}

\item{transform_fun}{A function to transform the fluorescence values. Default is \code{function(x) { asinh(x / 1e3) }}.
The reported MFIs are calculated as the median of the UNtransformed values. Transformation is only used to cluster the negative and positive populations.}

\item{relevant_columns}{A character vector specifying the relevant columns for multi-staining. Default is c("FITC-A", "PE-A", "ECD-A", "PC5.5-A", "PC7-A", "APC-A", "AF700-A", "AA750-A", "PB-A", "KrO-A").}
}
\value{
A data frame with the extracted MFIs. E.g.:
\preformatted{
 feature negative positive negative.sd positive.sd
<chr>      <dbl>    <dbl>     <dbl>      <dbl>
1 FITC-A      530.   58567.     576.      576.
2 PE-A        526.  149506.     511.      511.
}
}
\description{
Extract Multi-Stain Median Fluorescence Intensity (MFI)
This function extracts the median fluorescence intensity (MFI) from multi-stained FCS files.
}
