% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_mfi.R
\name{extract_relevant_mfis_singlestain}
\alias{extract_relevant_mfis_singlestain}
\title{Extract Single Stain Median Fluorescence Intensity (MFI)
This function extracts the median fluorescence intensity (MFI) from single-stain FCS files.}
\usage{
extract_relevant_mfis_singlestain(
  loaded_fcs_singlestain,
  transform_fun = function(x) {
     x
 },
  seed = 42
)
}
\arguments{
\item{loaded_fcs_singlestain}{A list of cytosets containing the loaded FCS files.}

\item{transform_fun}{A function to transform the fluorescence values. Default is \code{function(x) { asinh(x / 1e3) }}.
The reported MFIs are calculated as the median of the UNtransformed values. Transformation is only used to cluster the negative and positive populations.}
}
\value{
A data frame with the extracted MFIs. E.g.:
\preformatted{
  feature negative positive unstained
 <chr>      <dbl>    <dbl>     <dbl>
1 FITC-A      530.   58567.     576.
2 PE-A        526.  149506.     511.
}
}
\description{
Extract Single Stain Median Fluorescence Intensity (MFI)
This function extracts the median fluorescence intensity (MFI) from single-stain FCS files.
}
