% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_singlestain_mfi.R
\name{extract_singlestain_mfi_wrapper}
\alias{extract_singlestain_mfi_wrapper}
\title{Extract Single Stain Median Fluorescence Intensity (MFI)
This function extracts the median fluorescence intensity (MFI) from single-stain FCS files.}
\usage{
extract_singlestain_mfi_wrapper(
  loaded_fcs,
  transform_fun = function(x) {
     asinh(x/1000)
 },
  relevant_columns,
  ...
)
}
\arguments{
\item{loaded_fcs}{A list of cytosets containing the loaded FCS files.}

\item{transform_fun}{A function to transform the fluorescence values. Default is \code{function(x) { asinh(x / 1e3) }}.
The reported MFIs are calculated as the median of the UNtransformed values. Transformation is
only used to cluster the negative and positive populations.}
}
\value{
A data frame with the extracted MFIs. E.g.:
\preformatted{
# A tibble: 10 x 4
   feature negative positive unstained
   <chr>      <dbl>    <dbl>     <dbl>
 1 FITC-A      530.   58567.     576.
 2 PE-A        526.  149506.     511.
 3 ECD-A       453.  107584.     454.
 4 PC5.5-A     233.  242867.     166.
 5 PC7-A       176.  195307.     138.
 6 APC-A       136.  105036.     122.
 7 AF700-A     131.   28408.      97.2
 8 AA750-A     324.   75247.     296.
 9 PB-A        469.   13426.     442.
10 KrO-A       442.    4473.     444.
}
}
\description{
Extract Single Stain Median Fluorescence Intensity (MFI)
This function extracts the median fluorescence intensity (MFI) from single-stain FCS files.
}
