% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_mfi_quantile.R
\name{extract_singlestain_quantilefun}
\alias{extract_singlestain_quantilefun}
\title{Extract Quantile Approximation Functions for Single-Stain Flow Cytometry Data}
\usage{
extract_singlestain_quantilefun(
  loaded_fcs,
  n_quantiles = 100,
  quantiles = NULL,
  transform_fun = function(x) asinh(x/1000),
  approxmethod = function(...) {
     approxfun(..., rule = 2)
 },
  relevant_columns,
  ...
)
}
\arguments{
\item{loaded_fcs}{A named list of loaded FCS objects, where each element is a cytoset. Usually created by
\code{cytobench:::load_mfi_files}.}

\item{n_quantiles}{Integer.
Number of quantiles to use if \code{quantiles} is not provided. Default is 100.}

\item{quantiles}{Numeric vector.
Specific quantile probabilities to use (between 0 and 1). If \code{NULL}, defaults to evenly spaced quantiles.}

\item{transform_fun}{Function.
Transformation to apply to expression values. Only used for clustering the (then) transformed values.
Default is \code{asinh(x / 1e3)}.}

\item{approxmethod}{Function.
Method to approximate the quantile function (e.g., \code{approxfun}). Default is \code{approxfun}.}

\item{relevant_columns}{Character vector.
Names of marker columns to compute quantiles for.}

\item{...}{Additional arguments passed to the approximation method.}
}
\value{
A named list of lists, where each element corresponds to a sample and contains a quantile approximation
function for the relevant marker channel.
}
\description{
This function processes a list of loaded FCS files containing single-stain samples,
identifies the relevant marker channel, applies a transformation to the expression values,
separates positive and negative populations using k-means clustering, and computes
quantile approximation functions for the positive population.
}
\details{
For each sample, the function:
\itemize{
\item Identifies the non-empty marker channel.
\item Applies the specified transformation to the expression values.
\item Uses k-means clustering to separate positive and negative populations.
\item Restricts analysis to the positive population.
\item Computes quantile approximation functions for the specified marker columns.
}
Samples with no non-empty channels are skipped. Samples with multiple non-empty channels result in an error.
}
