% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSOM_is.outlier.R
\name{flowSOM_is.outlier}
\alias{flowSOM_is.outlier}
\title{Is a cell an outlier?}
\usage{
flowSOM_is.outlier(fsom, madAllowed = 4, fsomReference = NULL, channels = NULL)
}
\arguments{
\item{fsom}{FlowSOM object}

\item{madAllowed}{Number of median absolute deviations allowed. Default = 4.}

\item{fsomReference}{FlowSOM object to use as reference. If NULL (default),
the original fsom object is used.}

\item{channels}{If channels are given, return the result of FlowSOM::TestOutliers()}
}
\description{
Test if any cells are too far from their cluster centers.
This code is based on the FlowSOM package, FlowSOM::TestOutliers().
Here I do the exact same thing as in the FlowSOM package, but I return if the cell is an outlier or not.

In contrast, the FlowSOM package returns summaries of that, OR a per-channel outlier assignment.
This channelwise outlier assignment is different from the per-default reported outliers in the package itself.
}
\examples{
ff_example <- example_processed()
fsom <- do_flowsom_TESTING(ff_example)
outliers <- flowSOM_is.outlier(fsom)
}
\keyword{flowsom}
