% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSOM_optimal.R
\name{flowSOM_optimal}
\alias{flowSOM_optimal}
\title{flowSOM clustering with optimal number of clusters}
\usage{
flowSOM_optimal(
  fs_train,
  colsToUse,
  transform = TRUE,
  transformList = NULL,
  outdir = NULL,
  seed = 3711283,
  scale = TRUE,
  maxMeta = 50,
  ...
)
}
\arguments{
\item{fs_train}{The training set of the cytometry data.}

\item{colsToUse}{The columns to use for clustering.}

\item{transform}{Whether to transform the data using transformList}

\item{transformList}{The list of transformations to apply to the data. E.g.

fc_transforms <- sapply(colsToUse, function(cn_x) {
flowCore::arcsinhTransform(
a = 0,
b = 1 / pre_rescale_asinh[\link{"navios"}][\link{cn_x}],
c = 0
)
}, simplify = FALSE)
fc_transformlist_asinh <- flowCore::transformList(names(fc_transforms), fc_transforms)}

\item{outdir}{The directory to save the FlowSOM results.}

\item{seed}{The seed for the random number generator.}

\item{scale}{Whether to scale the data.}

\item{maxMeta}{The maximal number of metaclusters. If NULL, the number of clusters is FIXED by nClus.}

\item{...}{Further parameters to FlowSOM}
}
\description{
Clusters the cells using the FlowSOM algorithm with the optimal number of clusters (determined by ConsensusClusterPlus), by setting maxMeta.
}
