% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gate_cells.R
\name{gate_cells}
\alias{gate_cells}
\title{Gate Cells Based on Compensated FCS and Gating Set Files}
\usage{
gate_cells(
  flowset,
  gatingset,
  gatename = "/Singlets/CD45+/CD3+",
  verbose = TRUE,
  inplace = FALSE
)
}
\arguments{
\item{gatename}{Character or Numeric. Name or index of the gate to extract data from. Default is "/Singlets/CD45+/CD3+".}

\item{verbose}{Logical. If TRUE, prints progress messages. Default is TRUE.}

\item{dir_fcs_compensated}{Character. Directory containing compensated FCS files. Note that any compensation in these files is NOT applied.}

\item{dir_gatingsets_compensated}{Character. Directory containing compensated gating set files.
The gating set files are matched to each FCS file based on the sID (sampleID) from FCS and Gating \emph{directory} name.}

\item{pattern_files}{Character. Pattern to match FCS files. Default is "panel".}

\item{outdir}{Character. Output directory to save results. Default is NA, then no output is saved.}

\item{filename_ungated_into}{Character. String to replace "\emph{ungated}" in the filename. Default is "\emph{cd3}".}

\item{exclude_pops}{Character vector. Populations to exclude from the final counts. Default is "/Singlets/Lymphocytes", "/Singlets/Monocytes", "/Singlets/Granulocytes".}
}
\value{
A data.table containing the counts and percentages of gated populations. Usually you would use "outdir" as a side effect to save the data.
}
\description{
This function processes compensated FCS files and their corresponding gating set files to gate cells based on specified criteria.
}
\examples{
\dontrun{
gate_cells(
    dir_fcs_compensated = "path/to/compensated/fcs",
    dir_gatingsets_compensated = "path/to/gatingsets",
    outdir = "path/to/output"
)
}
}
