% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threedots.R
\name{GeomScattermore}
\alias{GeomScattermore}
\title{GeomScattermore}
\usage{
GeomScattermore(...)
}
\arguments{
\item{...}{Arguments passed to the \code{GeomScattermore} function.}
}
\value{
The result of the \code{GeomScattermore} function.
}
\description{
This function is a wrapper for the \code{GeomScattermore} function from the \code{scattermore} package.
}
\keyword{re-export}
