% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_rescale.R
\name{map_rescale}
\alias{map_rescale}
\title{Convenience function for rescaling multiple samples}
\usage{
map_rescale(
  sample_dt_list,
  mfi_dt_list,
  mapping_mfi_to_sample,
  verbose = TRUE,
  missing_feature = "center_median",
  known_missing_features = c("FS INT", "FS TOF", "SS INT", "TIME"),
  feature_unified_dict = feature_unified_dict_default(),
  sample_identifier_column = "fileX",
  ...
)
}
\arguments{
\item{sample_dt_list}{A named list of data.tables. Each data.table contains the cytometry data for one sample.}

\item{mfi_dt_list}{A named list of data.tables. Each data.table contains the MFI data for one sample.}

\item{mapping_mfi_to_sample}{A data.table with two columns: "fileX" and "mfi_file".
The "fileX" column contains the sample names and the "mfi_file" column contains the
MFI (file) names used to rescale that sample.}

\item{verbose}{Whether to print progress messages.}

\item{missing_feature}{If within the MFIs a feature is missing, this feature is rescaled according to the
value of this parameter. See \code{\link{rescale}} for details.}

\item{known_missing_features}{See \code{\link{rescale}} for details.}

\item{feature_unified_dict}{See \code{\link{rescale}} for details.}

\item{...}{Further arguments to \code{\link{rescale}}.}
}
\description{
Multiple samples can be rescaled at once using this function. It is necessary to
supply a "map" of
\itemize{
\item which MFI file belongs to which sample. One MFI can belong to multiple samples but also for a sample multiple MFIs can exist.
}
}
