% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfis_from_excel.R
\name{mfis_from_excel}
\alias{mfis_from_excel}
\title{Extract MFIs from Excel file}
\usage{
mfis_from_excel(
  excel_file = "Pre_Arcsinh_Median_FI.xlsx",
  positive_mfi_colname = "MFI POSITIVE POPULATION",
  negative_mfi_colname = c(extern = "MFI NEGATIVE POPULATION  UNSTAINED")
)
}
\arguments{
\item{excel_file}{The excel file must have a very specific format with the following columns:
EXP
CD3 LOT
MACHINE
CHANNEL
DYE
Name of the data file containing the data set
MFI FSC
MFI SSC
MFI POSITIVE POPULATION
MFI NEGATIVE POPULATION (GG)
MFI NEGATIVE POPULATION  UNSTAINED

Every single staining sample should be delimited by exactly one blank row.
The two machines "MFI NAVIOS" and "MFI EX" should be named sheets.}

\item{positive_mfi_colname}{Column name of the excel sheets with the positive MFI}

\item{negative_mfi_colname}{Column name of the excel sheets with the negative MFI.}
}
\value{
List ("intern", "extern") of matrices with MFIs per sample, looking like
this:
$intern
$intern$Align_01.Lot_A.NAVIOS
FL1 CD3 FITC FL2 CD3 PE FL3 CD3 ECD FL4 CD3 PC5.5 FL5 CD3 PC7 FL6 CD3 APC FL7 CD3 AF700 FL8 CD3 AA750 FL9 CD3 PB FL10 CD3 KrO
MFI POSITIVE POPULATION          37488.82   131657.6   139639.11     790976.38    451150.3    54247.44      37088.07      24695.75   27504.88     10440.54
MFI NEGATIVE POPULATION (GG)       219.82      298.0      437.83        615.24       364.9       66.68        164.55         98.35     824.71       792.06
}
\description{
Extract MFIs from Excel file
}
\examples{
# See tests
}
