% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_markers_pairwise.R
\name{plot_markers_pairwise}
\alias{plot_markers_pairwise}
\title{Plot all pairwise marker combinations with given cofactors}
\usage{
plot_markers_pairwise(
  ff_df,
  cofactor_namedvec,
  special_cofactor_list = list(),
  transform_fun = asinh,
  transform_fun_name = "asinh",
  verbose = FALSE,
  engine = c("ggplot", "base"),
  n_cells = Inf,
  title_global = NULL,
  modelines = TRUE,
  geom = c("hex", "points", "pointdensity"),
  bins = 50,
  diag_plot = FALSE,
  debugplots = FALSE,
  axis_full_labels = TRUE,
  count_transform = function(x) log10(x + 1),
  add_ggplot_elements = list(),
  ...
)
}
\arguments{
\item{ff_df}{A \code{flowFrame} or data.frame like  element containing the cytometry data.}

\item{cofactor_namedvec}{A named numeric vector providing default cofactors per marker. If not given, defaults to 1500 for all column names in \code{ff_df}.}

\item{special_cofactor_list}{A named list with overrides for specific marker combinations.
Each name should be \code{"markerX_markerY"}, with values being a
numeric vector of length 2: \code{c(cofactor_x, cofactor_y)}.}

\item{transform_fun}{A function used to transform intensity values (default: \code{asinh}).}

\item{transform_fun_name}{Character label for the transformation function (used in axis labels).}

\item{verbose}{Logical. If TRUE, prints progress messages to the console.}

\item{engine}{Character. One of \code{"ggplot"} (default) or \code{"base"} - determines the plotting engine used.}

\item{n_cells}{Number of cells/rows to downsample randomly (default: \code{Inf} = use all).}

\item{title_global}{Character. An optional global title for the entire plot.}

\item{modelines}{Logical. If TRUE, adds red lines at the modes of each marker in the scatterplots.}

\item{geom}{Character.
One of \code{"hex"} (default), \code{"points"}, or \code{"pointdensity"} -
determines the geometry used.
Speed concerns:
\itemize{
\item For engine ggplot2, hex is usually the fastest.
\item Fastest overall is \code{"points"} with engine \code{"base"}.
}}

\item{bins}{\code{engine = "ggplot"}. Number of bins (resolution) in each 2D plot (default: 50). (hex geom)}

\item{diag_plot}{Logical, \code{engine = "ggplot"}. Whether to include labeled marker names along the diagonal (default: FALSE).}

\item{debugplots}{Logical, \code{engine = "ggplot"}. If TRUE, plots text placeholders instead of actual data - useful for layout testing.}

\item{axis_full_labels}{Logical, \code{engine = "ggplot"}. If TRUE, axis labels include marker names and cofactor formulas.}

\item{count_transform}{Function to transform hexbin/pointdensity counts or density estimates (e.g. \code{log10(count + 1)}).}

\item{add_ggplot_elements}{\code{engine = "ggplot"}. A list of ggplot layers to add to each panel (e.g. \code{list(ggplot2::theme_minimal())}).}
}
\value{
A \code{patchwork} object: a composite grid of all marker pairwise comparisons.
}
\description{
Generates a matrix of 2D-density plots for all marker combinations in a flow cytometry dataset.
Axes can be transformed and scaled using marker-specific cofactors. This is particularly useful
for visual QC of compensation, cofactor choices and signal separation.
}
\examples{
fs <- simulate_fs(
    n_samples = 2,
    flowcore = TRUE,
    ncells = 250,
    columns = c("FL1-A", "FL2-A", "FL3-A")
)
p0 <- plot_markers_pairwise(fs[[1]])
print(p0)
p1 <- plot_markers_pairwise(
    fs[[1]],
    cofactor_namedvec = c("FL1-A" = 5, "FL2-A" = 10, "FL3-A" = 15),
    special_cofactor_list = list("FL1-A_FL2-A" = c(3, 6), "FL2-A_FL3-A" = c(4, 8)),
    transform_fun = function(x) log10(x + 1),
    transform_fun_name = "log10p1"
)
print(p1)
p2 <- plot_markers_pairwise(
    fs[[1]],
    cofactor_namedvec = c("FL1-A" = 5, "FL2-A" = 10, "FL3-A" = 15),
    special_cofactor_list = list("FL1-A_FL2-A" = c(3, 6), "FL2-A_FL3-A" = c(4, 8)),
    geom = "points",
    bins = 100,
    diag_plot = TRUE,
    debugplots = FALSE,
    axis_full_labels = TRUE,
    n_cells = 1000
)
print(p2)
p3 <- plot_markers_pairwise(
    fs[[1]],
    cofactor_namedvec = c("FL1-A" = 5, "FL2-A" = 10, "FL3-A" = 15),
    special_cofactor_list = list("FL1-A_FL2-A" = c(3, 6), "FL2-A_FL3-A" = c(4, 8)),
    geom = "pointdensity",
    bins = 100,
    diag_plot = TRUE,
    debugplots = FALSE,
    axis_full_labels = TRUE,
    n_cells = 1000,
    add_ggplot_elements = list(ggplot2::theme_minimal())
)
print(p3)
}
\keyword{cytometry}
