% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threedots.R
\name{REEXPORT-ns-triplecolon}
\alias{REEXPORT-ns-triplecolon}
\alias{\%:::\%}
\title{Access Internal Functions in a Package Namespace}
\usage{
pkg \%:::\% fun
}
\arguments{
\item{pkg}{A character string. The name of the package.}

\item{fun}{A character string. The name of the internal function to access.}
}
\value{
The function object retrieved from the specified package namespace.
}
\description{
This infix operator provides access to non-exported (internal) functions
from a package's namespace, similar to using \code{pkg:::fun} in base R.
Gotten from:
https://stat.ethz.ch/pipermail/r-devel/2013-August/067210.html
}
\details{
I need this in for the functions defined in the this very script "threedots.R".
This is a workaround to access internal functions that are not exported by the package.
It allows you to use the syntax \code{pkg \%:::\% fun} to retrieve the function \code{fun} from the package \code{pkg}.
}
\examples{
\dontrun{
# Access the internal .unitize function from the grid package
"grid" \%:::\% ".unitize"
}
}
\keyword{re-export}
