% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_extracted.R
\name{rescale_extracted}
\alias{rescale_extracted}
\title{Rescale a sample using a named list of extracted MFIs}
\usage{
rescale_extracted(
  sample_to_rescale,
  extracted_mfi,
  column_negative = c("negative", "unstained"),
  column_positive = "positive",
  missing_feature = c("minmax", "center_median"),
  inplace_datatable = FALSE,
  scale_column_fun = scale_column_relative,
  ...
)
}
\arguments{
\item{sample_to_rescale}{data.table of the sample to rescale}

\item{column_negative}{The column name of the negative population}

\item{column_positive}{The column name of the positive population}

\item{missing_feature}{How should missing features (NA) be handled?}

\item{inplace_datatable}{Should the data.table be modified in place?}

\item{scale_column_fun}{Function to scale a column}

\item{...}{Additional arguments passed to the scale_column_fun}

\item{extracted_mfi_df}{The extracted MFIs from extract_singlestain_mfi. The column "feature" must
contain the names of the columns to rescale.
By default, if a column of sample_to_rescale is not found in the extracted_mfi_namedlist,
the column is rescaled using the minmax method.}
}
\description{
This function rescales all columns of a sample using the extracted MFIs from extract_singlestain_mfi.
}
