% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_named.R
\name{rescale_named}
\alias{rescale_named}
\title{Rescale a sample using a named list of extracted MFIs}
\usage{
rescale_named(
  sample_to_rescale,
  extracted_mfi_namedlist,
  missing_feature = c("minmax", "center_median"),
  inplace_datatable = FALSE,
  scale_column_fun = scale_column_relative,
  ...
)
}
\arguments{
\item{sample_to_rescale}{data.table of the sample to rescale}

\item{extracted_mfi_namedlist}{The extracted MFIs as list.
In contrast to rescale(), this is a named list of positive and negative MFIs
where the names are the columns of the sample_to_rescale.
If a column should be rescaled as "missing feature",
the corresponding list element should be NA.
By default, if a column of sample_to_rescale is not found in the extracted_mfi_namedlist,
the column is rescaled using the minmax method.}

\item{missing_feature}{How should missing features (NA) be handled?}

\item{inplace_datatable}{Should the data.table be modified in place?}

\item{scale_column_fun}{Function to scale a column}

\item{...}{Additional arguments passed to the scale_column_fun}
}
\description{
This function rescales all columns of a sample using a named list of extracted MFIs.
}
