% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_named_quantiles.R
\name{rescale_named_quantiles}
\alias{rescale_named_quantiles}
\title{Rescale a sample using a named list of extracted MFIs}
\usage{
rescale_named_quantiles(
  sample_to_rescale,
  extracted_mfi_namedlist,
  reference_marker,
  missing_feature = c("minmax", "center_median", NA),
  inplace_datatable = FALSE,
  scale_column_fun = scale_column_minmaxQ,
  ...
)
}
\arguments{
\item{sample_to_rescale}{data.table of the sample to rescale}

\item{extracted_mfi_namedlist}{The extracted MFIs as list of lists:
In contrast to rescale_named(), this is a named list of LISTs with the names "positive" and
"negative" MFIs or have no name.
names(extracted_mfi_namedlist) should be the names of the columns in sample_to_rescale.
names(extracted_mfi_namedlist\link{1}) should either be "positive", "negative" or have no name
If it has no name, I assume that this column should be regarded as a "missing feature".

By default, if a column of sample_to_rescale is not found in the extracted_mfi_namedlist,
the column is rescaled using the minmax method.}

\item{missing_feature}{How should missing features (NA) be handled?}

\item{inplace_datatable}{Should the data.table be modified in place?}

\item{scale_column_fun}{Function to scale a column}

\item{...}{Additional arguments passed to the scale_column_fun}
}
\description{
This function rescales all columns of a sample using a named list of extracted MFIs.
}
