% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_column_identity.R
\name{scale_column_identity}
\alias{scale_column_identity}
\title{Identity Scaling Function (No Operation)}
\usage{
scale_column_identity(sample_to_rescale, ...)
}
\arguments{
\item{sample_to_rescale}{A \code{data.table} (or compatible object) to be returned unchanged.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{sample_to_rescale} without modification.
}
\description{
This function performs no scaling on the data. It is used as a placeholder or for debugging
when no transformation is required. It invisibly returns the input unchanged.
}
\examples{
ff_dt <- simulate_ff(columns = c("FL1", "FL2", "FL3"), flowcore = FALSE)
dt2 <- data.table::data.table(ff_dt)
scale_column_identity(
    sample_to_rescale = dt2,
    scaling_values = c(10, 100),
    colX = "FL1",
    subtract_bg = TRUE
)
scale_column_identity(
    sample_to_rescale = dt2
)
}
\keyword{relativisation}
