% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_column_minmax.R
\name{scale_column_minmax}
\alias{scale_column_minmax}
\title{Min-Max Scaling for a Single Column}
\usage{
scale_column_minmax(
  sample_to_rescale,
  scaling_values,
  colX,
  subtract_bg = TRUE
)
}
\arguments{
\item{sample_to_rescale}{A \code{data.table} containing the data to be rescaled.}

\item{scaling_values}{A numeric vector of length 1 or 2:
\itemize{
\item If length 2: used as (min, max) for scaling.
\item If length 1: interpreted as median and centered by subtracting it (see below).
}}

\item{colX}{The name of the column (as a string) to scale.}

\item{subtract_bg}{Logical. If TRUE, the minimum value (\code{low}) is subtracted from the data.
Defaults to TRUE. Ignored when centering on the median.}
}
\value{
Invisibly returns the modified \code{sample_to_rescale}.
}
\description{
This function scales a specified column of a data.table using provided
scaling values, neither usually nor necessarily to \link{0, 1}.
If a single value is provided, then only one median given which should be centered around 0.
The transformation is performed in-place.
}
\details{
If \code{length(scaling_values) == 1}, the column is centered using:
\deqn{x' = x - scaling\_values}
This is equivalent to subtracting the median.

If \code{length(scaling_values) == 2}, then traditional min-max scaling is applied:
\deqn{x' = (x - low) / (high - low)}

The operation modifies the input \code{sample_to_rescale} in place.
}
