% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_column_minmaxQ.R
\name{scale_column_minmaxQ}
\alias{scale_column_minmaxQ}
\title{Rescale a Column Using Quantile-Based Min-Max Normalization}
\usage{
scale_column_minmaxQ(
  sample_to_rescale,
  scaling_values,
  colX,
  subtract_bg = TRUE
)
}
\arguments{
\item{sample_to_rescale}{A data.table containing the data to be rescaled. Must include \code{reference_quantiles}.}

\item{scaling_values}{Either a numeric vector of length 2 (\code{c(min, max)}), or a named list
with elements \code{"negative"} and \code{"positive"}, each being a numeric value or a function
returning a value when applied to \code{reference_quantiles}.
The scaling function should take a quantile value and return the corresponding high value.}

\item{colX}{Character string specifying the column in \code{sample_to_rescale} to be rescaled.}

\item{subtract_bg}{Logical; if \code{TRUE}, the lower bound (negative control) is subtracted before scaling.}
}
\value{
Invisibly returns the modified \code{sample_to_rescale}, with the column \code{colX} scaled.
}
\description{
This function rescales a column in a data.table using min-max scaling, where the
minimum and maximum are either fixed numeric values or functions that return values
based on \code{reference_quantiles}, which must be present in the \code{sample_to_rescale}.
It supports standard two-point scaling and more flexible quantile-based corrections.
}
