% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_staining_extract_mfi_named.R
\name{single_staining_extract_mfi_named}
\alias{single_staining_extract_mfi_named}
\title{Extract Median Fluorescence Intensities from single-stained samples}
\usage{
single_staining_extract_mfi_named(
  single_stainings,
  sample_unstained = NULL,
  ignore_features = c("FS INT", "FS TOF", "SS INT", "TIME"),
  minimum_diff = 5000
)
}
\arguments{
\item{single_stainings}{DESCRIPTION.}

\item{sample_unstained}{DESCRIPTION.}

\item{relevant_difference}{DESCRIPTION.}

\item{minimum_fraction_per_cluster}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
Extract
}
\examples{
# ADD_EXAMPLES_HERE

}
