% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threedots.R
\name{spill2txt}
\alias{spill2txt}
\title{spill2txt}
\usage{
spill2txt(...)
}
\arguments{
\item{...}{Arguments passed to the \code{spill2txt} function.}
}
\value{
The result of the \code{spill2txt} function.
}
\description{
This function is a wrapper for the \code{spill2txt} function from the \code{flowCore} package.
}
\keyword{re-export}
