% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threedots.R
\name{string_to_spill}
\alias{string_to_spill}
\title{string_to_spill}
\usage{
string_to_spill(...)
}
\arguments{
\item{...}{Arguments passed to the \code{string_to_spill} function.}
}
\value{
The result of the \code{string_to_spill} function.
}
\description{
This function is a wrapper for the \code{string_to_spill} function from the \code{flowCore} package.
}
