% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample_fcs.R
\name{subsample_fcs}
\alias{subsample_fcs}
\title{Subsample FCS File}
\usage{
subsample_fcs(
  file_x,
  n_cells = 10000,
  seed = 427764,
  outdir = paste0("res/a01/02_subsampled/random_", n_cells),
  indir = "res/a04/01_gated",
  verbose = TRUE
)
}
\arguments{
\item{file_x}{Character. Path to the input FCS file.}

\item{n_cells}{Integer. Number of cells to subsample. Default is 10,000.}

\item{seed}{Integer. Seed for random number generation to ensure reproducibility. Default is 427764.}

\item{outdir}{Character. Directory where the subsampled FCS file will be saved. Default is "res/a01/02_subsampled/random_<n_cells>".}

\item{indir}{Character. Directory where the original FCS file is located. Default is "res/a04/01_gated".}

\item{verbose}{Logical. If TRUE, prints progress messages. Default is TRUE.}
}
\value{
None. The function writes the subsampled FCS file to the specified output directory.
}
\description{
This function reads an FCS file, subsamples it to a specified number of cells,
and writes the subsampled data to a new file.
}
\examples{
ff_dt <- simulate_ff(columns = c("FL1", "FL2", "FL3"), flowcore = TRUE)
tmpdir <- local_tempdir_time()
flowCore::write.FCS(ff_dt, file.path(tmpdir, "test.fcs"))
list.files(tmpdir)
# Run the subsampling function
subsample_fcs(
    file_x = file.path(tmpdir, "test.fcs"),
    n_cells = 100,
    seed = 42,
    outdir = file.path(tmpdir, "subsampled"),
    indir = tmpdir,
    verbose = FALSE
)
ff <- flowCore::read.FCS(file.path(tmpdir, "subsampled", "test.fcs"))
if (nrow(ff) != 100) {
    stop("Subsampling did not return the expected number of cells.")
}
unlink(tmpdir, recursive = TRUE) # Clean up the temporary directory

}
\keyword{cytometry}
