% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample_fcs.R
\name{subsample_ff}
\alias{subsample_ff}
\title{Subsample or Upsample a FlowFrame}
\usage{
subsample_ff(flowframe, n_cells = 10000, seed = 427764)
}
\arguments{
\item{flowframe}{A FlowFrame object containing the data to be subsampled or upsampled.}

\item{n_cells}{An integer specifying the number of cells to sample. Default is 10,000.}

\item{seed}{An integer specifying the seed for random number generation. Default is 427764.}
}
\value{
A FlowFrame object containing the subsampled or upsampled data.
}
\description{
This function subsamples or upsamples a given FlowFrame object to a specified number of cells.
If the number of cells in the FlowFrame is less than the specified number, the function
replicate the cells until the specified number is reached.
E.g. given 100 cells and n_cells = 5000, the function will replicate each cell of the
100 cells exactly 50 times.
}
\examples{
# Assuming `ff` is a FlowFrame object with flow cytometry data
subsampled_ff <- subsample_ff(ff, n_cells = 5000, seed = 12345)

}
