% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweep_matrix_col.R
\name{sweep_matrix_col}
\alias{sweep_matrix_col}
\title{Apply a Function to Each Column Using Column-wise Statistics}
\usage{
sweep_matrix_col(x, stats, fun = "-")
}
\arguments{
\item{x}{A numeric matrix. Each column will be modified using the corresponding value in \code{stats}.}

\item{stats}{A numeric vector with length equal to \code{ncol(x)}. These are the per-column statistics to apply with \code{fun}.}

\item{fun}{A function or a string naming a function (e.g., \code{"-"}, \code{"+"}, \code{"*"}, \code{"/"}).
Defaults to subtraction (\code{"-"}).}
}
\value{
A numeric matrix of the same dimensions as \code{x}, with \code{fun(x[, i], stats[i])} applied column-wise.
}
\description{
Applies a function (e.g., subtraction) between each column of a matrix \code{x} and a
corresponding statistic value provided in \code{stats}.
This is equivalent to applying \code{sweep()} with \code{MARGIN = 2}, but faster.
}
\examples{
mat <- matrix(1:9, nrow = 3)
stats <- c(1, 2, 3)
sweep_matrix_col(mat, stats, fun = "-")

n <- 1e3
n_colums <- 1e2
n_rows <- n / n_colums
m <- matrix(rnorm(n_colums), nrow = n_rows, ncol = n_colums)
stats <- rnorm(n_colums)
# microbenchmark::microbenchmark(
#     v1 = sweep(m, 2, stats, "-"),
#     v2 = sweep_matrix_col(m, stats, "-"),
#     times = 100,
#     check = "equal"
# )
}
