% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempdir_time.R
\name{tempdir_time}
\alias{tempdir_time}
\title{Create Timestamped Temporary Directory}
\usage{
tempdir_time()
}
\value{
A character string containing the full path to the newly created timestamped directory.
The directory is created immediately on disk.
}
\description{
Creates a new temporary directory by appending a timestamp to the default \code{tempdir()}.
This is useful for generating isolated working directories in pipelines or scripts.
}
\examples{
tmp <- tempdir_time()
list.files(tmp) # should be empty

}
