% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threedots.R
\name{UpdateDerivedValues}
\alias{UpdateDerivedValues}
\title{UpdateDerivedValues}
\usage{
UpdateDerivedValues(...)
}
\arguments{
\item{...}{Arguments passed to the \code{UpdateDerivedValues} function.}
}
\value{
The result of the \code{UpdateDerivedValues} function.
}
\description{
This function is a wrapper for the \code{UpdateDerivedValues} function from the \code{FlowSOM} package.
}
\keyword{re-export}
