% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmv_helper_insert_compensations.R
\name{cmv_helper_insert_compensations}
\alias{cmv_helper_insert_compensations}
\title{Insert multiple compensations into a fcs file}
\usage{
cmv_helper_insert_compensations(
  fcs_filename,
  named_compensation_files = list(manual = NA, DCtubes_auto = NA, singlestain_auto = NA,
    singlestain_manual = NA),
  outdir = file.path("res", "CT_UnappliedCompensations"),
  sub_filename = c("intermediate/CT", ""),
  spillover_keyword = "$SPILLOVER",
  verbose = TRUE
)
}
\arguments{
\item{fcs_filename}{The path to the fcs file to insert the compensations into}

\item{named_compensation_files}{A named list of files containing the compensation files. The names are the compensation types.}

\item{outdir}{The directory to save the fcs files with the inserted compensations. If NA, the function returns the flowFrame object with the inserted compensations.}

\item{sub_filename}{A vector of two strings to substitute in the fcs_filename. The first string is substituted by the second string. The final files are saved in \code{file.path(outdir, sub(sub_filename[1], sub_filename[2], fcs_filename))}}

\item{spillover_keyword}{The keyword in the fcs file containing the original spillover matrix.}

\item{verbose}{Whether to print out the path of the saved fcs file}
}
\description{
This function is a helper function to INSERT the compensation into the fcs files of the CMV data from
Glehr et al. \link{TO BE SUBMITTED}.
}
