% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_marker_mfi_list.R
\name{extract_marker_mfi_list}
\alias{extract_marker_mfi_list}
\title{Extract Marker MFIs as Named List for Rescaling}
\usage{
extract_marker_mfi_list(
  sample_to_rescale,
  extracted_mfi,
  column_negative = c("negative", "unstained"),
  column_positive = "positive"
)
}
\arguments{
\item{sample_to_rescale}{A \code{flowFrame} or other object with marker (column) names to be rescaled.}

\item{extracted_mfi}{A \code{tibble} (or coercible to one) with at least three columns:
\describe{
\item{\code{feature}}{Character. Names of the features (e.g., \code{"FITC-A"}).}
\item{\code{negative}}{Numeric. MFI values for negative/unlabeled control.}
\item{\code{positive}}{Numeric. MFI values for positive/labeled control.}
}}

\item{column_negative}{Character vector. Column name as  negative control.
Defaults to \code{c("negative", "unstained")}. First element is used.}

\item{column_positive}{Character. Column name used as positive control. Defaults to \code{"positive"}. First element is used.}
}
\value{
A named list. Each entry corresponds to a marker (column) in \code{sample_to_rescale},
containing a numeric vector of length two with names \code{"negative"} and \code{"positive"}.
Missing markers are filled with \code{NA}.
}
\description{
This function extracts median fluorescence intensities (MFIs) for specified features
from a tibble, returning a named list that maps each feature (e.g., "FITC-A") to a
named numeric vector of its negative and positive control values. Features present
in \code{sample_to_rescale} but missing from \code{extracted_mfi} are included with \code{NA}.
}
\examples{
set.seed(42)
fs_ss <- simulate_cd3()
tmpdir <- local_tempdir_time()
flowCore::write.flowSet(fs_ss, tmpdir)

extracted_mfis_singlestain <- extract_mfi(
    tmpdir,
    regex_singlestain = "(-(CD3-.*)|(none))\\\\.fcs$"
)
extract_marker_mfi_list(
    sample_to_rescale = fs_ss[["sample0_12-panel"]],
    extracted_mfi = extracted_mfis_singlestain,
    column_negative = "negative",
    column_positive = "positive"
)
}
\keyword{internal}
\keyword{relativisation}
