% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_mfi.R
\name{extract_singlestain_mfi}
\alias{extract_singlestain_mfi}
\title{Extract Single Stain Median Fluorescence Intensity (MFI)}
\usage{
extract_singlestain_mfi(
  fcs_dir = "data-raw/s001",
  regex_singlestain = "(-(CD3-.*)|(none))\\\\.fcs$",
  transform = function(x) {
     asinh(x/1000)
 },
  gating_set_file = NULL,
  gate_extract = NULL
)
}
\arguments{
\item{fcs_dir}{A character string specifying the directory containing the FCS files. Default is "data-raw/s001".}

\item{regex_singlestain}{A regular expression pattern to identify single-stain FCS files. Default is "(-(CD3-.*)|(none))\\.fcs$".}

\item{transform}{As \code{transform_fun} in \code{extract_mfi}.}

\item{gating_set_file}{A gating set file or object to apply to the FCS files. Default is NULL.
Usually you want to gate lymphocytes before extracting the MFIs. If a character string is provided,
the gating set is loaded from the file. If a (GatingSet) object is provided, it is used directly.}

\item{gate_extract}{A specific gate to extract from the gating set. Default is NULL.}
}
\description{
Extract Single Stain Median Fluorescence Intensity (MFI)
}
\examples{
set.seed(42)
fs_ss <- simulate_cd3()
tmpdir <- local_tempdir_time()
flowCore::write.flowSet(fs_ss, tmpdir)
loaded_fcs <- load_mfi_files(
    fcs_dir = tmpdir,
    regex_singlestain = "(-(CD3-.*)|(none))\\\\.fcs$",
    gating_set_file = NULL,
    gate_extract = NULL
)
extract_relevant_mfis_singlestain(loaded_fcs)
}
\keyword{relativisation}
