% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSOM_newMetacluster.R
\name{flowSOM_newMetacluster}
\alias{flowSOM_newMetacluster}
\title{Predict new cells using a flowsom result}
\usage{
flowSOM_newMetacluster(
  flowsom_result,
  clustered_df,
  n_metacluster = NULL,
  update_flowsom = TRUE
)
}
\arguments{
\item{flowsom_result}{Result of flowSOM_optimal() or flowSOM()}

\item{clustered_df}{\enumerate{
\item The data.frame including the clustered cells. The column "cluster" must be present and
should be the cluster assignment from the FlowSOM ("flowsom_result").
\item Alternatively: The data.frame including the number of clustered cells. The columns "cluster_1", "cluster_2", ..., "cluster_n".
The function will then calculate the metaClusters for each of the clusters and return the number of cells per metaCluster.
}}

\item{n_metacluster}{If not NULL, the number of metaclusters will be changed to this number in flowsom_result.}
}
\value{
A list with two elements:
\enumerate{
\item "cells_clusters_from_train": The data.frame with the new metaClusters. If the input was a data.frame with the number of cells per cluster,
the output will be a data.frame with the number of cells per metaCluster. (Identical to 2))
\item "ncells_per_x": A data.frame in wide format with the counts of cells per specified grouping.
}
}
\description{
Predict the (meta-) clusters for all cells using a flowsom result.
}
