% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSOM_optimal.R
\name{flowSOM_optimal}
\alias{flowSOM_optimal}
\title{FlowSOM Clustering with Optimal Number of MetaClusters}
\usage{
flowSOM_optimal(
  fs_train,
  colsToUse,
  transform = TRUE,
  transformList = NULL,
  outdir = NULL,
  seed = 3711283,
  scale = TRUE,
  maxMeta = 50,
  ...
)
}
\arguments{
\item{fs_train}{A \code{flowSet} or \code{cytoset} representing the cytometry training data.}

\item{colsToUse}{Character vector. Column names (channels/markers) to include in clustering.}

\item{transform}{Logical. Whether to apply \code{transformList} before clustering (default \code{TRUE}).}

\item{transformList}{A \code{transformList} (typically from \code{flowCore::transformList}) defining the transformations to apply.
If \code{transform = TRUE}, this will be applied during clustering.E.g.

fc_transforms <- sapply(colsToUse, function(cn_x) {
flowCore::arcsinhTransform(
a = 0,
b = 1 / pre_rescale_asinh[\link{"navios"}][\link{cn_x}],
c = 0
)
}, simplify = FALSE)
fc_transformlist_asinh <- flowCore::transformList(names(fc_transforms), fc_transforms)}

\item{outdir}{Optional. Directory path to save the FlowSOM result object and associated PDF plots.}

\item{seed}{Integer. Random seed for reproducibility. Default is \code{3711283}.}

\item{scale}{Logical. Whether to scale data (default \code{TRUE}). Scaling is applied prior to SOM training.}

\item{maxMeta}{Integer. Maximum number of meta-clusters to consider. If \code{NULL}, defaults to a FIXED number via \code{nClus} argument.}

\item{...}{Additional arguments passed to \code{FlowSOM::FlowSOM()}.}
}
\value{
An object of class \code{"flowSOM_optimal"}, which includes:
\describe{
\item{\code{fs_res_train}}{The FlowSOM result object with meta-clustering annotations.}
\item{...}{Additional elements returned by \code{flowSOM_predict()}, including predicted cluster and meta-cluster labels.}
}
}
\description{
Runs the FlowSOM algorithm on a cytometry dataset and determines an optimal number
of meta-clusters using \code{ConsensusClusterPlus}. Automatically generates and optionally
saves plots and the clustering result. Can also return per-cell cluster predictions.
}
