% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSOM_predict.R
\name{flowSOM_predict}
\alias{flowSOM_predict}
\title{Predict new cells using a flowsom result}
\usage{
flowSOM_predict(flowsom_result, flowset, colsToUse)
}
\arguments{
\item{flowsom_result}{Result of flowSOM_optimal() or flowSOM()}

\item{flowset}{The flowset whose cells should be assigned to the clusters from flowsom_result}

\item{colsToUse}{The columns to use for clustering.}
}
\description{
Predict the (meta-) clusters for all cells using a flowsom result.
}
