% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_cells_per_x.R
\name{n_cells_per_x}
\alias{n_cells_per_x}
\title{Calculate the number of cells per specified grouping}
\usage{
n_cells_per_x(
  dt,
  which = c("cluster", "metaCluster"),
  id_cols = c("sample"),
  which_levels = NULL
)
}
\arguments{
\item{dt}{A data.table containing the data.}

\item{which}{A character vector specifying the columns to group by. Default is c("cluster", "metaCluster").}

\item{id_cols}{A character vector specifying the ID columns. Default is c("sample").}

\item{which_levels}{A list specifying the levels for each grouping column. Default is NULL.
If NULL, the levels are inferred from the data by:
\enumerate{
\item Converting the unique values of the column to numeric.
\item Taking the maximum value (max_level_numeric)
\item Setting the levels to 1:max_level_numeric
}}
}
\value{
A data frame in wide format with the counts of cells per specified grouping.
}
\description{
This function calculates the number of cells per specified grouping (e.g., cluster or metaCluster)
and returns a wide-format data frame with the counts.
}
\examples{
# Example usage:
dt <- data.table::data.table(
    sample = as.character(rep(1:3, each = 4)),
    cluster = rep(1:2, times = 6),
    metaCluster = rep(1:4, times = 3)
)
n_cells_per_x(dt)
}
