% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_filename.R
\name{parse_filename}
\alias{parse_filename}
\title{Parse filenames into parts}
\usage{
parse_filename(
  fn,
  cn = c("study", "patientID", "tvt", "sample_nr", "processing", "exported_gate",
    "downsampling_method", "n_cells", "machine", "dye")
)
}
\arguments{
\item{fn}{Filenames, usually a vector of multiple strings}

\item{cn}{The colnames the filename(s) should be split into}
}
\description{
Extracts defined parts of the filename from the filename string
}
