% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_markers_pairwise.R
\name{plot_markers_pairwise}
\alias{plot_markers_pairwise}
\title{Plot all pairwise marker combinations with given cofactors}
\usage{
plot_markers_pairwise(
  ff,
  cofactor_namedvec,
  special_cofactor_list,
  transform_fun = asinh,
  transform_fun_name = "asinh",
  geom = c("hex", "points", "pointdensity"),
  bins = 50,
  diag_plot = FALSE,
  debugplots = FALSE,
  axis_full_labels = TRUE,
  n_cells = Inf,
  count_transform = function(x) {
     log10(x + 1)
 },
  add_ggplot_elements = list()
)
}
\arguments{
\item{ff}{A flowFrame object}

\item{cofactor_namedvec}{A named vector with the cofactors for each marker. The names of the vector should be the same as the marker names in the flowFrame object.}

\item{special_cofactor_list}{A named list with special cofactors for certain marker combinations. The names of the list should be the marker combinations, separated by an underscore. The values should be a vector of two values, the first for the x-axis marker and the second for the y-axis marker.}

\item{transform_fun}{The transformation function to apply to the data. Default is asinh.}

\item{transform_fun_name}{The name of the transformation function. Default is "asinh".}

\item{bins}{The number of bins for the hexbin plot}

\item{diag_plot}{Whether to include the diagonal plots}

\item{debugplots}{Whether to include debug plots (No cells, only text to show the layout)}

\item{axis_full_labels}{Whether to include the cofactors in the axis labels}

\item{n_cells}{The number of cells to plot. Default is Inf, which plots all cells.}

\item{count_transform}{A function to transform the count values. Default is the identity function.}

\item{add_ggplot_elements}{A list of ggplot elements to add to the plot. Default is an empty list.}
}
\description{
Plots all pairwise marker combinations with given cofactors. The cofactors are used to scale the markers before plotting. The plot is a hexbin plot of the asinh transformed values of the markers. The cofactors are used to scale the markers before plotting.
}
\examples{
pdf("removeme.pdf", width = 60, height = 50)
print(
    plot_combinations(
        current_sample_gated[[1]],
        pre_rescale_asinh_sscheck[[device_x]],
        special_asinh[[device_x]],
        debugplots = TRUE
    )
)
print(
    plot_combinations(
        current_sample_gated[[1]],
        pre_rescale_asinh_sscheck[[device_x]],
        special_asinh[[device_x]],
        diag_plot = FALSE,
        debugplots = TRUE
    )
)
dev.off()

}
