% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_extracted.R
\name{rescale_extracted}
\alias{rescale_extracted}
\title{Rescale Sample Using Extracted MFIs}
\usage{
rescale_extracted(
  sample_to_rescale,
  extracted_mfi,
  column_negative = c("negative", "unstained"),
  column_positive = "positive",
  missing_feature = c("minmax", "center_median"),
  inplace_datatable = FALSE,
  scale_column_fun = scale_column_minmax,
  ...
)
}
\arguments{
\item{sample_to_rescale}{A \code{flowFrame} or other object with marker (column) names to be rescaled.}

\item{extracted_mfi}{A \code{tibble} (or coercible to one) with at least three columns:
\describe{
\item{\code{feature}}{Character. Names of the features (e.g., \code{"FITC-A"}).}
\item{\code{negative}}{Numeric. MFI values for negative/unlabeled control.}
\item{\code{positive}}{Numeric. MFI values for positive/labeled control.}
}}

\item{column_negative}{Character vector. Column name as  negative control.
Defaults to \code{c("negative", "unstained")}. First element is used.}

\item{column_positive}{Character. Column name used as positive control. Defaults to \code{"positive"}. First element is used.}

\item{missing_feature}{Character. How to handle missing MFI entries (i.e., where list entry is \code{NA}).
Options are:
\itemize{
\item \code{"minmax"} (default): Rescale using column-wise min/max.
\item \code{"center_median"}: Shift column so that median is centered.
\item \code{NA}: Do not rescale columns with missing MFI values.
}}

\item{inplace_datatable}{Logical. If \code{TRUE}, modifies \code{sample_to_rescale} in place. If \code{FALSE}, returns a new copy. Default is \code{FALSE}.}

\item{scale_column_fun}{Function. The scaling function to apply per column. Default is \code{scale_column_minmax()}.
This function must accept at least \code{sample_to_rescale} (data.table), \code{scaling_values} (vector of numbers), and \code{colX} (character, a column of \code{sample_to_rescale}).}

\item{...}{Additional arguments passed to \code{scale_column_fun}.}
}
\description{
Rescales all features (columns) in a cytometry sample using median fluorescence intensities (MFIs) extracted from controls.
}
\keyword{relativisation}
