% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample_fcs.R
\name{subsample_fcs}
\alias{subsample_fcs}
\title{Subsample FCS File}
\usage{
subsample_fcs(
  file_x,
  n_cells = 10000,
  seed = 427764,
  outdir = paste0("res/a01/02_subsampled/random_", n_cells),
  indir = "res/a04/01_gated",
  verbose = TRUE
)
}
\arguments{
\item{file_x}{Character. Path to the input FCS file.}

\item{n_cells}{Integer. Number of cells to subsample. Default is 10,000.}

\item{seed}{Integer. Seed for random number generation to ensure reproducibility. Default is 427764.}

\item{outdir}{Character. Directory where the subsampled FCS file will be saved. Default is "res/a01/02_subsampled/random_<n_cells>".}

\item{indir}{Character. Directory where the original FCS file is located. Default is "res/a04/01_gated".}

\item{verbose}{Logical. If TRUE, prints progress messages. Default is TRUE.}
}
\value{
None. The function writes the subsampled FCS file to the specified output directory.
}
\description{
This function reads an FCS file, subsamples it to a specified number of cells,
and writes the subsampled data to a new file.
}
\examples{
\dontrun{
subsample_fcs(
    "path/to/input.fcs",
    n_cells = 5000,
    seed = 12345,
    outdir = "output/directory",
    indir = "input/directory",
    verbose = TRUE
)
}

}
\keyword{cytometry}
