% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_memory_FCS.R
\name{write_memory_FCS}
\alias{write_memory_FCS}
\title{Write FCS File to Memory}
\usage{
write_memory_FCS(ff)
}
\arguments{
\item{ff}{A flowFrame object to be written to a file.}
}
\value{
A character string representing the path to the written FCS file.
}
\description{
This function attempts to write a flowFrame object to a temporary file in memory.
If writing to the default memory location fails, it creates a new temporary directory
and writes the file there.
}
\examples{
\dontrun{
  library(flowCore)
  ff <- read.FCS("path/to/your/file.fcs")
  fcs_path <- write_memory.FCS(ff)
  print(fcs_path)
}
}
